refactor(api): move user anonymized details for admin serializer
in src/identity-access-management
