+++ api/src/identity-access-management/infrastructure/serializers/jsonapi/user-anonymized-details-for-admin.serializer.js
